requirements.txt
astroid==1.5.3
avro-python3==1.8.2
certifi==2017.7.27.1
chardet==3.0.4
confluent-kafka==0.11.0
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==1.7.4
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
